Sprint 16 MVVM First state
