+++ app/src/main/java/com/practicum/playlistmaker/TrackItemAdapter.kt
@@ -8,7 +8,8 @@ import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import com.google.gson.Gson
-import com.practicum.playlistmaker.playerActivity.presentation.PlayerActivity
+import com.practicum.playlistmaker.player.domain.entities.Track
+import com.practicum.playlistmaker.player.presentation.PlayerActivity
 
 var isClickAllowed = true
