+++ app/src/main/java/com/practicum/playlistmaker/player/domain/dao/TrackIntentDAO.kt
@@ -1,0 +1,7 @@
+package com.practicum.playlistmaker.player.domain.dao
+
+import com.practicum.playlistmaker.player.domain.entities.Track
+
+interface TrackIntentDAO {
+    fun getTrack():Track
+}
