+++ app/src/main/java/com/practicum/playlistmaker/player/domain/interfaces/MediaPlayerPrepare.kt
@@ -1,4 +1,4 @@
-package com.practicum.playlistmaker.playerActivity.domain.interfaces
+package com.practicum.playlistmaker.player.domain.interfaces
 
 interface MediaPlayerPrepare {
     fun onPrepared()
