+++ app/src/main/java/com/practicum/playlistmaker/player/domain/interfaces/TrackInteractor.kt
@@ -1,0 +1,11 @@
+package com.practicum.playlistmaker.player.domain.interfaces
+
+import com.practicum.playlistmaker.player.domain.entities.MediaPlayerState
+import com.practicum.playlistmaker.player.domain.entities.Track
+
+interface TrackInteractor {
+    fun getTrack(): Track
+    fun startPlayer(): MediaPlayerState
+    fun pausePlayer(): MediaPlayerState
+    fun playbackControl(): MediaPlayerState
+}
