+++ app/src/main/java/com/practicum/playlistmaker/player/domain/interfaces/TrackRepository.kt
@@ -1,0 +1,14 @@
+package com.practicum.playlistmaker.player.domain.interfaces
+
+import com.practicum.playlistmaker.player.domain.entities.MediaPlayerState
+import com.practicum.playlistmaker.player.domain.entities.Track
+
+interface TrackRepository {
+    fun getTrack(): Track
+
+    fun preparePlayer(mediaPlayerPreparator:MediaPlayerPrepare)
+
+    fun startPlayer():MediaPlayerState
+
+    fun pausePlayer():MediaPlayerState
+}
