--- app/src/main/java/com/practicum/playlistmaker/playerActivity/domain/interfaces/TrackInteractor.kt
@@ -1,11 +1,0 @@
-package com.practicum.playlistmaker.playerActivity.domain.interfaces
-
-import com.practicum.playlistmaker.playerActivity.domain.entities.MediaPlayerState
-import com.practicum.playlistmaker.playerActivity.domain.entities.Track
-
-interface TrackInteractor {
-    fun getTrack(): Track
-    fun startPlayer(): MediaPlayerState
-    fun pausePlayer(): MediaPlayerState
-    fun playbackControl(): MediaPlayerState
-}
