+++ app/src/main/res/layout/activity_player.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.practicum.playlistmaker.playerActivity.presentation.PlayerActivity">
+    tools:context="com.practicum.playlistmaker.player.presentation.PlayerActivity">
     <!-- Guideline start-->
     <androidx.constraintlayout.widget.Guideline
         android:id="@+id/guideline_start"
